更新了旋转编码器控制电机的文件，但是只能加不能减，不知道怎么解决

+++ 5-3旋转编码器驱动电机/Hardware/Motor.c
@@ -18,7 +18,7 @@ void Motor_Init(void)
 	PWM_Init();
 }
 
-void Motor_SetSpeed(int8_t speed)
+void Motor_SetSpeed(int16_t speed)
 {
 	if(speed>=0)
 	{
